Update issue 3
Add the remaining test cases. Now all functioning test cases have 
been added. Missing are 

TEST15 : This one does not appear to work. Needs a close look 
         why 
TEST17 : Needs a file from ASDEX Upgrade which is not available 
         Can perhaps be modified. 

All test cases have been added, but have not been looked at 
in detail. This is necessary to ensure that the original version 
of the code passes the (even older) test cases. 

Some of the output of the testcases needs to be rewritten. 
Sometimes very small numbers are written out that depend on the 
machine accuracy. These tests do not necessarily pass when 
changing compiler, compiler switches or hardware. 

Nevertheless, with the testcases in place one can start 
changing the source, while making sure of not breaking the 
implementation. 



git-svn-id: https://neoart.googlecode.com/svn/trunk@27 eddb0e7f-f83d-3e0f-2fd2-5d76d495fea8